+++ src/secrets/totp.rs
@@ -513,6 +513,7 @@ impl Totp<'_> {
 #[cfg(test)]
 mod tests {
     #![allow(clippy::panic)]
+    #![allow(deprecated)]
 
     use secrecy::SecretString;